+++ resources/views/components/feedback.blade.php
@@ -13,7 +13,7 @@
                     <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
                 </div>
                 <div class="modal-body">
-                    <x-ig::form action="{{ route('feedback.send') }}" id="feedbackForm" class="editable-skip" :recaptcha="false">
+                    <x-ig::form action="{{ route('feedback.send') }}" id="feedbackForm" class="editable-skip">
                         <x-ig::input
                             type="text"
                             name="subject"