+++ src/Http/Controllers/FeedbackController.php
@@ -5,13 +5,15 @@ namespace InternetGuru\LaravelFeedback\Http\Controllers;
 use Illuminate\Http\Request;
 use Illuminate\Routing\Controller;
 use Illuminate\Support\Facades\Mail;
+use InternetGuru\LaravelCommon\Contracts\ReCaptchaInterface;
 use InternetGuru\LaravelCommon\Support\Helpers;
 use InternetGuru\LaravelFeedback\Mail\FeedbackMail;
 
 class FeedbackController extends Controller
 {
-    public function send(Request $request)
+    public function send(Request $request, ReCaptchaInterface $recaptcha)
     {
+        $recaptcha->validate($request);
         $validated = $request->validate([
             'subject' => 'required|string|max:255',
             'message' => 'required|string',